py_toolkits/requirements.txt
mysqlclient==2.1.1
pymongo==4.1.1
PySocks==1.7.1
